scripts/youtube_comment_replier/requirements.txt
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
openai==1.59.8
